requirements.txt
# Shared domain models and database
biz2bricks-core @ git+https://github.com/cwijay/biz_to_bricks_core_v1.git@main

# LangGraph & LangChain
langgraph>=1.0.4
langchain>=1.2.0
langchain-openai>=1.1.0
langchain-google-genai>=3.2.0
langchain-community>=0.4.1
google-genai>=1.53.0
python-dotenv>=1.0.0
# Pin cryptography for langgraph-api compatibility
cryptography>=42.0.0,<45.0.0

# LlamaIndex
llama-cloud-services>=0.6.88
# Pin llama-cloud for llama-index compatibility
llama-cloud>=0.1.35,<0.2.0

# PostgreSQL (Cloud SQL)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
cloud-sql-python-connector[asyncpg]>=1.5.0
nest-asyncio>=1.6.0
pgvector>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-multipart>=0.0.9

# Production Resilience
tenacity>=8.2.0

# Excel/Sheets Agent
pandas>=2.0.0
duckdb>=0.9.0
openpyxl>=3.0.0

# Google Cloud Storage
google-cloud-storage>=2.14.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0